Move Draw from Unit to Drawable
