+++ engine/src/cmd/drawable.h
@@ -119,7 +119,9 @@ public:
     void SetAniSpeed( float speed );
 
     //Uses GFX so only in Unit class
-    virtual void Draw( const Transformation &quat = identity_transformation, const Matrix &m = identity_matrix ) = 0;
+    virtual void Draw( const Transformation &quat = identity_transformation,
+                       const Matrix &m = identity_matrix );
+    virtual void DrawNow();
     virtual void DrawNow( const Matrix &m = identity_matrix, float lod = 1000000000 ) = 0;
     virtual std::string drawableGetName() = 0;
 
@@ -137,6 +139,9 @@ public:
 
     //Lights the shields, without applying damage or making the AI mad - useful for special effects
     void LightShields( const Vector &pnt, const Vector &normal, float amt, const GFXColor &color );
+
+    ///Holds temporary values for inter-function XML communication Saves deprecated restr info
+    Matrix WarpMatrix( const Matrix &ctm ) const;
 };
 
 Matrix* GetCumulativeTransformationMatrix(Unit *unit, const Matrix &parentMatrix, Matrix invview);
